+++ src/lapidary/render/elems/operation_function.py
@@ -15,7 +15,7 @@ from .response_body import response_type_name
 from ..module_path import ModulePath
 from ..type_ref import TypeRef, get_type_ref, GenericTypeRef, resolve_type_ref
 from ...openapi import model as openapi
-from ...openapi.model import LapisType
+from ...openapi.model import LapidaryModelType
 
 RESPONSE_BODY = 'response_body'
 
@@ -130,7 +130,7 @@ def get_response_types(op: openapi.Operation, module: ModulePath, resolve: Resol
             else:
                 name = response_type_name(op, resp_code)
                 resp_module = module / RESPONSE_BODY
-            if schema.lapis_model_type is LapisType.exception:
+            if schema.lapidary_model_type is LapidaryModelType.exception:
                 continue
             typ = resolve_type_ref(schema, resp_module, name, resolve)
             response_types.add(typ)